+++ scim2_tester/__init__.py
@@ -1,3 +1,5 @@
 from .checker import check_server
+from .utils import CheckResult
+from .utils import Status
 
-__all__ = ["check_server"]
+__all__ = ["check_server", "Status", "CheckResult"]