+++ scim2_tester/service_provider_config.py
@@ -1,9 +1,7 @@
 from typing import Tuple
 
-from httpx import HTTPError
-from pydantic import ValidationError
 from scim2_client import SCIMClient
-from scim2_models import Error
+from scim2_client import SCIMClientError
 from scim2_models import Resource
 from scim2_models import ServiceProviderConfig
 from utils import CheckResult
@@ -25,21 +23,8 @@ def check_service_provider_config_endpoint(
     """
 
     try:
-        response = scim.query(ServiceProviderConfig)
-    except HTTPError as exc:
-        return CheckResult(status=Status.ERROR, reason=str(exc)), None
-
-    except ValidationError as exc:
-        return (
-            CheckResult(
-                status=Status.ERROR,
-                reason=f"Could not validate the response payload: {exc}",
-                data=exc.response_payload,
-            ),
-            None,
-        )
-
-    if isinstance(response, Error):
-        return CheckResult(status=Status.ERROR, reason=response.detail), response
+        response = scim.query(ServiceProviderConfig, raise_scim_errors=True)
+    except SCIMClientError as exc:
+        return CheckResult(status=Status.ERROR, reason=str(exc), data=exc.source)
 
     return CheckResult(status=Status.SUCCESS), response